# Optimize barycentric interpolation