### Purpose of this PR

* Description of feature/change.

I noticed during benchmarking math operations that one of the Barycentric overloads was four times slower than the other.

Turns out the [slow version](https://github.com/opentk/opentk/blob/a2e1be1293cc7cefe8307bb404d0d92c82369317/src/OpenToolkit.Mathematics/Vector/Vector3.cs#L746) contains a naive implementation using operator overloads (leading to a number of unnecessary copies/allocations), while the [fast version](https://github.com/opentk/opentk/blob/a2e1be1293cc7cefe8307bb404d0d92c82369317/src/OpenToolkit.Mathematics/Vector/Vector3.cs#L764) has an optimized implementation that passes all parameters by reference.

This change replaces the naive implementations with a call to the optimized method instead.

* Which part of OpenTK does this affect (Math, OpenGL, Platform, Input, etc).

Math

* Links to screenshots, design docs, user docs, etc.

Se comments for impact of current approach.

### Testing status

* Explanation of what’s tested, how tested and existing or new automation tests.

Ran unit tests, and verified the Barycentric ones all passed. Note that the double variants of the vectors do not have existing test coverage.

### Comments

Benchmark of current implementations as of a2e1be1293cc7cefe8307bb404d0d92c82369317:

|                                                                                                           Method |       Runtime |       Mean |      Error |     StdDev |     Median | Ratio | RatioSD |
|----------------------------------------------------------------------------------------------------------------- |-------------- |-----------:|-----------:|-----------:|-----------:|------:|--------:|
| &#39;Vector3.BaryCentric(Vector3 a, Vector3 b, Vector3 c, float u, float v)&#39;                                 | .NET Core 3.1 |  29.595 ns |  0.0812 ns |  0.0720 ns |  29.596 ns |  0.99 |    0.00 |
| &#39;Vector3.BaryCentric(ref Vector3 a, ref Vector3 b, ref Vector3 c, float u, float v, out Vector3 result)&#39; | .NET Core 3.1 |   7.048 ns |  0.1418 ns |  0.1326 ns |   7.082 ns |  0.96 |    0.02 |