@cryovat 2020-02-25
Benchmark with change:

|                                                                                                   Method |       Runtime |      Mean |     Error |    StdDev | Ratio |
|--------------------------------------------------------------------------------------------------------- |-------------- |----------:|----------:|----------:|------:|
|                                 &#39;Vector3.BaryCentric(Vector3 a, Vector3 b, Vector3 c, float u, float v)&#39; | .NET Core 3.1 | 15.229 ns | 0.0645 ns | 0.0572 ns |  0.98 |
| &#39;Vector3.BaryCentric(ref Vector3 a, ref Vector3 b, ref Vector3 c, float u, float v, out Vector3 result)&#39; | .NET Core 3.1 |  7.106 ns | 0.0357 ns | 0.0334 ns |  1.04 |

I was honestly hoping for a bigger improvement, but 2x cost still beats 4x I guess...